% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_soo.R
\name{theme_soo}
\alias{theme_soo}
\title{A theme for gpplot2}
\usage{
theme_soo(fill = "transparent")
}
\arguments{
\item{object}{A ggplot2 object}
}
\value{
A customized plot with Soo's theme!
}
\description{
A theme for gpplot2
}
