% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_soo.R
\name{theme_soo}
\alias{theme_soo}
\title{A theme for gpplot2}
\usage{
theme_soo(base_family = "Avenir", title_size = 22,
  facet_text_size = 12, facet_fill = "#7fcdbb", facet_height = 0.3,
  fill = "transparent", colour = "transparent", title_position = 0.5,
  axis_title_size = 14, axis_tick_size = 12)
}
\arguments{
\item{base_family}{The font family to be used, defaults to Avenir}

\item{title_size}{Title size, defaults to 22}

\item{facet_text_size}{Facet text size, defaults to 12}

\item{facet_fill}{Color of facet rectangle, defaults to "#7fcdbb"}

\item{fill}{Background element rectangles, defaults to "transparent"}

\item{colour}{Colour option for key, defaults to "transparent"}

\item{title_position}{Title position, defaults to 0.5}

\item{axis_title_size}{Axis title size, defaults to 14}

\item{axis_tick_size}{Size of ticks on axes, defaults to 12}
}
\value{
A customized plot with Soo's theme!
}
\description{
A theme for gpplot2
}
